install browserify
