--- popup/css/popup.css
@@ -1,23 +1,0 @@
-html, body{
-  width: 250px;
-  height: 60px;
-  padding: 0px;
-  margin: 0px;
-}
-div.button{
-  color: #545454;
-  padding-left: 10px;
-  cursor: pointer;
-  line-height: 30px;
-  height: 30px;
-  font-size: 13px;
-}
-div.button:hover {
-  background-color: rgba(1, 117, 152, 0.2);
-}
-.button > .icon{
-  height: 15px;
-  font-size: 17px;
-  line-height: 15px;
-  vertical-align: middle;
-}
