--- popup/popup.html
@@ -1,20 +1,0 @@
-<!-- DOCTYPE html -->
-<html>
-  <head>
-    <meta charset="utf-8"/>
-    <link href="./css/popup.css" rel="stylesheet" type="text/css" />
-    <link href="./css/font-awesome.min.css" rel="stylesheet" type="text/css" />
-  </head>
-  <body>
-    <div id='capture' class='button'>
-      <i class="fa fa-crop icon"></i>
-    </div>
-    <div id='whole' class='button'>
-      <img src='/icons/browser.svg' class='icon'/>
-    </div>
-    <div id='selectElm' class='button'>
-      <i class="fa fa-search icon"></i>
-    </div>
-    <script src='popup.js'></script>
-  </body>
-</html>
