+++ script/branch2crx
@@ -5,7 +5,7 @@ var exec = require('child_process').execSync
 var token = process.env.GITHUB_TOKEN
 
 try {
-  exec('npm run pack')
+  exec('npm run build && npm run pack')
 } catch (e) {
   console.error('Error on packing')
   process.exit(-1)
