+++ src/statics/_locales/en/messages.json
@@ -1,0 +1,35 @@
+{
+  "uploadingFinishTitle": {
+    "message": "Complete!"
+  },
+  "uploadingFinishMessage": {
+    "message": "Gyazo link has been copied"
+  },
+  "captureTitle": {
+    "message": "Capturing..."
+  },
+  "captureMessage": {
+    "message": "Please wait"
+  },
+  "contextMenuImage": {
+    "message": "Capture this image"
+  },
+  "selectArea": {
+    "message": "Free Selection"
+  },
+  "topToBottom": {
+    "message": "Whole Page"
+  },
+  "selectElement": {
+    "message": "Element"
+  },
+  "captureWindow": {
+    "message": "Page"
+  },
+  "confirmReload": {
+    "message": "This extension requires you to reload tabs after install. Could you reload the current tab now?"
+  },
+  "welcomeMessage": {
+    "message": "Welcome! Please open a page in a new tab, and click the Gyazo button!"
+  }
+}
