+++ src/statics/_locales/ja/messages.json
@@ -1,0 +1,35 @@
+{
+  "uploadingFinishTitle": {
+    "message": "完了!"
+  },
+  "uploadingFinishMessage": {
+    "message": "Gyazoのリンクがコピーされました"
+  },
+  "captureTitle": {
+    "message": "キャプチャー中..."
+  },
+  "captureMessage": {
+    "message": "お待ちください"
+  },
+  "contextMenuImage": {
+    "message": "この画像をアップロード"
+  },
+  "selectArea": {
+    "message": "範囲選択"
+  },
+  "topToBottom": {
+    "message": "ページ全体"
+  },
+  "selectElement": {
+    "message": "部品"
+  },
+  "captureWindow": {
+    "message": "ページ"
+  },
+  "confirmReload": {
+    "message": "この拡張を使うためには一度タブをリロードする必要があります。今すぐリロードしますか？"
+  },
+  "welcomeMessage": {
+    "message": "ようこそ。キャプチャしたいウェブサイトを新しいタブで開いて、Gyazoボタンを押してみてください。"
+  }
+}
