+++ src/statics/fonts/icomoon.woff2
